[refactor]post, comment 생성 및 변경에 Custom Hook 적용
