+++ src/components/Layout/Container/PostsContainer.jsx
@@ -1,70 +1,31 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import PostContainer from './PostContainer';
 import { useDispatch, useSelector } from 'react-redux';
-import { __addPost, __getPost } from '../../../redux/modules/posts';
-import { useNavigate } from 'react-router-dom';
-import uuid from 'react-uuid';
+import { __getPost } from '../../../redux/modules/posts';
 import * as S from './PostsContainerStyle';
-import dayjs from 'dayjs';
 import blankProfile from '../../../images/blankProfile.webp';
-import { auth } from '../../../firebase';
+import { usePostCreate } from '../../../hooks/usePostCreate';
 
 const PostsContainer = () => {
   const dispatch = useDispatch();
-  const navigate = useNavigate();
   const { posts } = useSelector((state) => state.posts);
   const { user } = useSelector((state) => state.user);
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const [imgUrl, setImgUrl] = useState(blankProfile);
-  const [imgDownloadUrl, setImagDownloadUrl] = useState(null);
-  // const [imgUploaded, setImgUploaded] = useState(false);
-  // const [user, setUser] = useState('anonymous');
-  // console.log('2', imgDownloadUrl);
-
-  //task 추가 버튼
-  const onSubmitHandler = (e) => {
-    e.preventDefault();
-
-    if (title === '' || content === '') return; // 아무것도 입력하지 않았을 때 dispatch 하지 않음
-    // console.log('imgUrl값은?', imgUrl);
-    // console.log(user[0].invitedUid);
-    user.length > 0 //로그인 해야만 디스패치 되도록 조건 처리
-      ? dispatch(
-          __addPost({
-            userName: auth.currentUser
-              ? auth.currentUser.displayName
-              : user[0].email.split('@')[0],
-            created_at: dayjs().format('YYYY.MM.DD HH:mm:ss'),
-            id: uuid(),
-            title,
-            content,
-            isDone: false,
-            userId: user[0].id,
-            imgUrl: imgDownloadUrl ?? blankProfile,
-            invitedId: user[0].invitedUid ?? '',
-          })
-        )
-      : alert('로그인해주세요');
-
-    setTitle('');
-    setContent('');
-    navigate('/');
-  };
+  //post 생성 커스텀 훅 적용
+  const [
+    { title, content, imgUrl, imgDownloadUrl },
+    changeTitle,
+    changeContent,
+    onSubmitHandler,
+  ] = usePostCreate({
+    title: '',
+    content: '',
+    imgUrl: blankProfile,
+    imgDownloadUrl: null,
+  });
 
   useEffect(() => {
     dispatch(__getPost());
   }, [dispatch]);
-
-  // console.log('posts: ', posts);
-  useEffect(() => {
-    if (posts.length < 1) {
-      return;
-    }
-    posts.map((post) => setImgUrl(post.imgUrl));
-  }, [posts]);
-
-  // console.log(auth.currentUser?.photoURL);
 
   return (
     <S.CommentsWrap>
@@ -78,9 +39,7 @@ const PostsContainer = () => {
             className='title'
             type='text'
             value={title}
-            onChange={(e) => {
-              setTitle(e.target.value);
-            }}
+            onChange={changeTitle}
           />
           <S.TitleCommentText>내용</S.TitleCommentText>
 
@@ -89,9 +48,7 @@ const PostsContainer = () => {
             rows='10'
             type='text'
             value={content}
-            onChange={(e) => {
-              setContent(e.target.value);
-            }}
+            onChange={changeContent}
           />
 
           <S.AddBtn disabled={title === '' || content === '' ? true : false}>
